FormClaimEdit, click OK, there was an error about specified cast not valid.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3293 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/Clearinghouses.cs
@@ -203,7 +203,7 @@ namespace OpenDentBusiness{
 				RefreshCache();
 			}
 			if(payorID!="" && HList.ContainsKey(payorID)){
-				return (int)HList[payorID];
+				return (long)HList[payorID];
 			}
 			//payorID not found
 			Clearinghouse defaultCH=GetDefault();
